tender-agent/requirements.txt
playwright>=1.42.0
playwright-stealth>=1.0.6
fake-useragent>=1.4.0
aiohttp>=3.9.0
pydantic>=2.6.0
rich>=13.7.0
questionary>=2.0.1
pyyaml>=6.0.1
pandas>=2.1.0
aiosqlite>=0.20.0
openai>=1.30.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
